+++ src/datatable/tests/unit/assets/datatable-body-tests.js
@@ -419,16 +419,34 @@ suite.add(new Y.Test.Case({
                     thousandsSeparator: 't',
                     prefix: 'p',
                     suffix: 's'
+                }},
+                {key: 'lookup', formatter: 'lookup', lookupTable: [
+                    {value: undefined, text: 'unknown'},
+                    {value: 0, text: 'zero'},
+                    {value: 1, text: 'one'},
+                    {value: 2, text: 'two'},
+                    {value: 3, text: 'three'},
+                    {value: 4, text: 'four'}
+                ]},
+                {key: 'lookup1', formatter: 'lookup', lookupTable: {
+                    undefined: 'unknown',
+                    0: 'zero',
+                    1: 'one',
+                    2: 'two',
+                    3: 'three',
+                    4: 'four'
                 }}
 
             ],
             data: [
                 {
-                    a: 123.45, b: 123.45, button:'btn', 'boolean': true, 'date': this.dateFormatTime,
-                    localDate: this.dateFormatTime, localTime: this.dateFormatTime, localDateTime: this.dateFormatTime,
-                    email: 'me', link: 'site', linkSrc: 'there', number: 987654
+                    a: 123.45, b: 123.45, button:'btn', 'boolean': true, 'date': new Date(),
+                    localDate: new Date(), localTime: new Date(), localDateTime: new Date(),
+                    email: 'me', link: 'site', linkSrc: 'there', number: 987654,
+                    lookup: 1, lookup1: 1
                 },
-                {a: 6789,   b: 6789  , email: 'me', link: 'site',              'boolean': false },
+                {a: 6789,   b: 6789  , email: 'me', link: 'site',              'boolean': false,
+                    lookup:3, lookup1:3 },
                 {}
             ],
             currencyFormat: {
@@ -490,18 +508,18 @@ suite.add(new Y.Test.Case({
                 }
                 return val;
             };
-        
+
         Y.Assert.isTrue(node.hasClass('yui3-datatable-date'));
         Y.Assert.areEqual(getTestText(Y.Date.format(this.dateFormatTime)), node.getHTML());
-        
+
         node = dt.getCell([0,5]);
         Y.Assert.isTrue(node.hasClass('yui3-datatable-date'));
         Y.Assert.areEqual(getTestText(Y.Date.format(this.dateFormatTime,{format:'%x'})), node.getHTML());
-        
+
         node = dt.getCell([0,6]);
         Y.Assert.isTrue(node.hasClass('yui3-datatable-date'));
         Y.Assert.areEqual(getTestText(Y.Date.format(this.dateFormatTime,{format:'%X'})), node.getHTML());
-        
+
         node = dt.getCell([0,7]);
         Y.Assert.isTrue(node.hasClass('yui3-datatable-date'));
         Y.Assert.areEqual(getTestText(Y.Date.format(this.dateFormatTime,{format:'%c'})), node.getHTML());
@@ -545,6 +563,24 @@ suite.add(new Y.Test.Case({
         Y.Assert.areEqual('p987t654d00s', node.getHTML());
         node = dt.getCell([2, 0]);
         Y.Assert.areEqual('', node.getHTML());
+    },
+    "test lookup format": function () {
+        var dt = this.dt,
+            node = dt.getCell([0,11]);
+        Y.Assert.isTrue(node.hasClass('yui3-datatable-lookup'));
+        Y.Assert.areEqual('one', node.getHTML());
+        node = dt.getCell(node, 'below');
+        Y.Assert.areEqual('three', node.getHTML());
+        node = dt.getCell(node, 'below');
+        Y.Assert.areEqual('unknown', node.getHTML());
+
+        node = dt.getCell([0,12]);
+        Y.Assert.isTrue(node.hasClass('yui3-datatable-lookup'));
+        Y.Assert.areEqual('one', node.getHTML());
+        node = dt.getCell(node, 'below');
+        Y.Assert.areEqual('three', node.getHTML());
+        node = dt.getCell(node, 'below');
+        Y.Assert.areEqual('unknown', node.getHTML());
     }
 }));